Fix
Fix Spark App container.
Writing the results locally and to hdfs works.
Query 2 Works.
Made the test to see how the aggregate data are generated, no arithmetic mean is used but perhaps a weighted average but the weights for the country zones are not known. We need to ask.

TODO:
- query 3;
- query 4;
- write results to Redis;
- start Grafana...

Co-Authored-By: Luca Martorelli <92364829+MartorelliLuca@users.noreply.github.com>